components/object_store/requirements.txt
dalloriam>=0.4.11
google-cloud-error-reporting
google-cloud-storage
